Refactored Wiring Transform
Added many comments
