+++ src/test/scala/firrtlTests/WiringTests.scala
@@ -202,4 +202,146 @@ class WiringTests extends FirrtlFlatSpec {
     val retC = wiringPass.run(c)
     (parse(retC.serialize).serialize) should be (parse(check).serialize)
   }
+  "Two sources" should "fail" in {
+    val sinks = Map(("X"-> "pin"))
+    val sas = WiringInfo("A", "clk", sinks, "Top")
+    val input =
+      """circuit Top :
+        |  module Top :
+        |    input clk: Clock
+        |    inst a1 of A
+        |    a1.clk <= clk
+        |    inst a2 of A
+        |    a2.clk <= clk
+        |  module A :
+        |    input clk: Clock
+        |    inst x of X
+        |    x.clk <= clk
+        |  extmodule X :
+        |    input clk: Clock
+        |""".stripMargin
+    intercept[WiringException] {
+      val c = passes.foldLeft(parse(input)) {
+        (c: Circuit, p: Pass) => p.run(c)
+      }
+      val wiringPass = new Wiring(sas)
+      val retC = wiringPass.run(c)
+    }
+  }
+  "Wiring from A.clk to X, with 2 A's, and A as top" should "work" in {
+    val sinks = Map(("X"-> "pin"))
+    val sas = WiringInfo("A", "clk", sinks, "A")
+    val input =
+      """circuit Top :
+        |  module Top :
+        |    input clk: Clock
+        |    inst a1 of A
+        |    a1.clk <= clk
+        |    inst a2 of A
+        |    a2.clk <= clk
+        |  module A :
+        |    input clk: Clock
+        |    inst x of X
+        |    x.clk <= clk
+        |  extmodule X :
+        |    input clk: Clock
+        |""".stripMargin
+    val check =
+      """circuit Top :
+        |  module Top :
+        |    input clk: Clock
+        |    inst a1 of A
+        |    a1.clk <= clk
+        |    inst a2 of A
+        |    a2.clk <= clk
+        |  module A :
+        |    input clk: Clock
+        |    inst x of X
+        |    x.clk <= clk
+        |    wire clk_0: Clock
+        |    clk_0 <= clk
+        |    x.pin <= clk_0
+        |  extmodule X :
+        |    input clk: Clock
+        |    input pin: Clock
+        |""".stripMargin
+    val c = passes.foldLeft(parse(input)) {
+      (c: Circuit, p: Pass) => p.run(c)
+    }
+    val wiringPass = new Wiring(sas)
+    val retC = wiringPass.run(c)
+    (parse(retC.serialize).serialize) should be (parse(check).serialize)
+  }
+  "Wiring from A.clk to X, with 2 A's, and A as top, but Top instantiates X" should "error" in {
+    val sinks = Map(("X"-> "pin"))
+    val sas = WiringInfo("A", "clk", sinks, "A")
+    val input =
+      """circuit Top :
+        |  module Top :
+        |    input clk: Clock
+        |    inst a1 of A
+        |    a1.clk <= clk
+        |    inst a2 of A
+        |    a2.clk <= clk
+        |    inst x of X
+        |    x.clk <= clk
+        |  module A :
+        |    input clk: Clock
+        |    inst x of X
+        |    x.clk <= clk
+        |  extmodule X :
+        |    input clk: Clock
+        |""".stripMargin
+    intercept[WiringException] {
+      val c = passes.foldLeft(parse(input)) {
+        (c: Circuit, p: Pass) => p.run(c)
+      }
+      val wiringPass = new Wiring(sas)
+      val retC = wiringPass.run(c)
+    }
+  }
+  "Wiring from A.r[a] to X" should "work" in {
+    val sinks = Map(("X"-> "pin"))
+    val sas = WiringInfo("A", "r[a]", sinks, "A")
+    val input =
+      """circuit Top :
+        |  module Top :
+        |    input clk: Clock
+        |    inst a of A
+        |    a.clk <= clk
+        |  module A :
+        |    input clk: Clock
+        |    reg r: UInt<2>[5], clk
+        |    node a = UInt(5)
+        |    inst x of X
+        |    x.clk <= clk
+        |  extmodule X :
+        |    input clk: Clock
+        |""".stripMargin
+    val check =
+      """circuit Top :
+        |  module Top :
+        |    input clk: Clock
+        |    inst a of A
+        |    a.clk <= clk
+        |  module A :
+        |    input clk: Clock
+        |    reg r: UInt<2>[5], clk
+        |    node a = UInt(5)
+        |    inst x of X
+        |    x.clk <= clk
+        |    wire r_a: UInt<2>
+        |    r_a <= r[a]
+        |    x.pin <= r_a
+        |  extmodule X :
+        |    input clk: Clock
+        |    input pin: UInt<2>
+        |""".stripMargin
+    val c = passes.foldLeft(parse(input)) {
+      (c: Circuit, p: Pass) => p.run(c)
+    }
+    val wiringPass = new Wiring(sas)
+    val retC = wiringPass.run(c)
+    (parse(retC.serialize).serialize) should be (parse(check).serialize)
+  }
 }
